+++ src/effects/filter/mod.rs
@@ -1,2 +1,3 @@
 pub mod biquad;
 pub mod ladder_filter;
+pub mod state_variable;