+++ src/effects/utility/mod.rs
@@ -1,4 +1,7 @@
 pub mod gain;
-pub mod passthrough;
 pub mod offset;
+pub mod passthrough;
 pub mod dc_source;
+pub mod add;
+pub mod multiply;
+pub mod panner;